:zip:新增日志记录

+++ src/main/java/com/cindy/ocrdemo/dto/AuditCheckDto.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 @Data
 public class AuditCheckDto {
-    private Integer invoiceId;
+    private Long invoiceId;
     //   1表示已通过  2表示待修改  3表示已拒绝
     private Integer status;
     private String opinion;
